# docs/swagger.yaml
basePath: /
definitions:
...
  httpapi.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  httpapi.PipelineParamDefDataResponse:
    properties:
      data:
        $ref: '#/definitions/httpapi.PipelineParamDefResponse'
    type: object
  httpapi.PipelineParamDefListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/httpapi.PipelineParamDefResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  httpapi.PipelineParamDefResponse:
    properties:
      can_edit:
        type: boolean
      can_view:
        type: boolean
      created_at:
        type: string
      default_value:
        type: string
      description:
        type: string
      editable:
        type: boolean
      executor_param_name:
        type: string
      executor_type:
        type: string
      id:
        type: string
      param_key:
        type: string
      param_type:
        type: string
      pipeline_id:
        type: string
      raw_meta:
        type: string
      required:
        type: boolean
      single_select:
        type: boolean
      sort_no:
        type: integer
      source_from:
        type: string
      updated_at:
        type: string
      visible:
        type: boolean
    type: object
...
  httpapi.SyncPipelineParamDefsResponse:
    properties:
      data:
        $ref: '#/definitions/usecase.SyncPipelineParamDefsOutput'
    type: object
...
  httpapi.UpdatePipelineParamDefRequest:
    properties:
      param_key:
        type: string
    type: object
...
  usecase.SyncPipelineParamDefsOutput:
    properties:
      created:
        type: integer
      skipped:
        type: integer
      total:
        type: integer
      updated:
        type: integer
    type: object
...
info:
  contact: {}
  description: Internal deployment platform API.
  title: GOS API
  version: "1.0"
paths:
...
  /applications/{id}/pipeline-param-defs:
    get:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Binding type, default ci
        in: query
        name: binding_type
        type: string
      - description: Visible flag
        in: query
        name: visible
        type: boolean
      - description: Editable flag
        in: query
        name: editable
        type: boolean
      - description: Mapped platform param key
        in: query
        name: param_key
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.PipelineParamDefListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: List bound Jenkins pipeline param definitions by application
      tags:
      - pipeline-param-defs
  /jenkins/pipeline-param-defs/sync:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.SyncPipelineParamDefsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: Sync pipeline param definitions from Jenkins
      tags:
      - pipeline-param-defs
...
  /pipeline-param-defs/{id}:
    get:
      parameters:
      - description: Pipeline param definition ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.PipelineParamDefDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: Get pipeline param definition by ID
      tags:
      - pipeline-param-defs
    put:
      consumes:
      - application/json
      parameters:
      - description: Pipeline param definition ID
        in: path
        name: id
        required: true
        type: string
      - description: Update pipeline param definition request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httpapi.UpdatePipelineParamDefRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.PipelineParamDefDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: Update pipeline param definition mapping
      tags:
      - pipeline-param-defs
...
  /pipelines/{id}/param-defs:
    get:
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: string
      - description: Executor type
        in: query
        name: executor_type
        type: string
      - description: Visible flag
        in: query
        name: visible
        type: boolean
      - description: Editable flag
        in: query
        name: editable
        type: boolean
      - description: Mapped platform param key
        in: query
        name: param_key
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.PipelineParamDefListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: List pipeline param definitions
      tags:
      - pipeline-param-defs
...
schemes:
- http
- https
swagger: "2.0"